test: Extract the tests for TestDMGMPermissionChecksForPosts  into a separate file to avoid conflict
